change mask data type

+++ 0x13-bit_manipulation/4-clear_bit.c
@@ -7,7 +7,7 @@
 */
 int clear_bit(unsigned long int *n, unsigned int index)
 {
-	unsigned long int mask;
+	unsigned int mask;
 
 	mask = 1;
 	if (n == NULL)
